// fe/src/lib/api/generated/model/platformPublicStatusResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { PublicPlatformStatus } from './publicPlatformStatus';

export interface PlatformPublicStatusResponse {
  ends_at: string | null;
  message: string | null;
  starts_at: string | null;
  status: PublicPlatformStatus;
}